Refactored some routes and folders
