+++ src/App/App.css
@@ -25,3 +25,13 @@ h2 {
     flex: 1;
     background-color: #191716;
 }
+
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+}
+
+input[type=number] {
+    -moz-appearance:textfield;
+}
